Move multithreaded_compute.rs into hello-compute tests
Setup hello-compute tests to run during `cargo test`
