+++ Cargo.toml
@@ -62,3 +62,8 @@ futures = "0.3"
 #wgc = { version = "0.1.0", package = "wgpu-core", path = "../wgpu/wgpu-core" }
 #wgt = { version = "0.1.0", package = "wgpu-types", path = "../wgpu/wgpu-types" }
 #wgn = { version = "0.4.0", package = "wgpu-native", path = "../wgpu/wgpu-native" }
+
+[[example]]
+name="hello-compute"
+path="examples/hello-compute/main.rs"
+test = true
